requirements.txt
asgiref==3.3.1
Django==3.1.6
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.2
gunicorn==20.0.4
Pillow==8.1.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
django_heroku==0.3.1
dj_database_url==0.5.0